+++ bridge/ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Add an opt-in Kafka record envelope for JSON transformations
+* Allow configuring Kafka consumer `auto.offset.reset`
 
 ## Version 1.99.0
 * Include the Kafka partition in consumer idempotency keys